Fixed review findings

+++ test/integration/api_access.py
@@ -1,5 +1,4 @@
-import os
-from typing import Iterable, List
+from typing import Iterable
 from contextlib import contextmanager
 from datetime import datetime
 
@@ -20,30 +19,27 @@ def timestamp() -> str:
     return f'{datetime.now().timestamp():.0f}'
 
 
-class OpenApiTestee:
-    def __init__(self):
-        self._client = OpenApiTestee.create_client()
-        self._account_id = os.environ["SAAS_ACCOUNT_ID"]
+def create_saas_client(
+        host: str,
+        pat: str,
+        raise_on_unexpected_status: bool = True,
+) -> openapi.AuthenticatedClient:
+    return openapi.AuthenticatedClient(
+        base_url=host,
+        token=pat,
+        raise_on_unexpected_status = raise_on_unexpected_status,
+    )
 
-    @classmethod
-    def create_client(cls, raise_on_unexpected_status: bool = True):
-        return openapi.AuthenticatedClient(
-            base_url=os.environ["SAAS_HOST"],
-            token=os.environ["SAAS_PAT"],
-            raise_on_unexpected_status = raise_on_unexpected_status,
-        )
 
-    def _get_client(self, ignore_failures: bool = False):
-        return (
-            OpenApiTestee.create_client(False)
-            if ignore_failures
-            else self._client
-        )
+class OpenApiAccess:
+    def __init__(self, client: openapi.Client, account_id: str):
+        self._client = client
+        self._account_id = account_id
 
-    def create_database(self) -> openapi.models.create_database.CreateDatabase:
+    def create_database(self, cluster_size: str = "XS") -> openapi.models.create_database.CreateDatabase:
         cluster_spec = openapi.models.CreateCluster(
             name="my-cluster",
-            size="XS",
+            size=cluster_size,
         )
         return create_database.sync(
             self._account_id,
@@ -56,12 +52,17 @@ class OpenApiTestee:
             )
         )
 
+    @contextmanager
+    def _ignore_failures(self, ignore: bool = False):
+        before = self._client.raise_on_unexpected_status
+        self._client.raise_on_unexpected_status = not ignore
+        yield self._client
+        self._client.raise_on_unexpected_status = before
+
     def delete_database(self, database_id: str, ignore_failures=False):
-        return delete_database.sync_detailed(
-            self._account_id,
-            database_id,
-            client=self._get_client(ignore_failures),
-        )
+        with self._ignore_failures(ignore_failures) as client:
+            return delete_database.sync_detailed(
+                self._account_id, database_id, client=client)
 
     def list_database_ids(self) -> Iterable[str]:
         dbs = list_databases.sync(self._account_id, client=self._client)
@@ -73,12 +74,13 @@ class OpenApiTestee:
             keep: bool = False,
             ignore_delete_failure: bool = False,
     ):
+        db = None
         try:
             db = self.create_database()
             yield db
         finally:
-            if not keep:
-                self.delete_database( db.id, ignore_delete_failure)
+            if not keep and db:
+                self.delete_database(db.id, ignore_delete_failure)
 
     def list_allowed_ip_ids(self) -> Iterable[openapi.models.allowed_ip.AllowedIP]:
         ips = list_allowed_i_ps.sync(
@@ -105,11 +107,9 @@ class OpenApiTestee:
         )
 
     def delete_allowed_ip(self, id: str, ignore_failures=False):
-        return delete_allowed_ip.sync_detailed(
-            self._account_id,
-            id,
-            client=self._get_client(ignore_failures),
-        )
+        with self._ignore_failures(ignore_failures) as client:
+            return delete_allowed_ip.sync_detailed(
+                self._account_id, id, client=client)
 
     @contextmanager
     def allowed_ip(
@@ -118,9 +118,10 @@ class OpenApiTestee:
             keep: bool = False,
             ignore_delete_failure: bool = False,
     ):
+        ip = None
         try:
             ip = self.add_allowed_ip(cidr_ip)
             yield ip
         finally:
-            if not keep:
-                self.delete_allowed_ip(id, ignore_delete_failure)
+            if not keep and ip:
+                self.delete_allowed_ip(ip.id, ignore_delete_failure)
